Fix debug messages format strings
CID: 452498, 452450, 452451, 452449, 452453, 452461, 452462, 452464, 452474, 452477, 452483, 452484, 452485, 452488, 452492, 452496, 452500, 452505, 452512, 452460, 452473, 452487, 452501, 452443, 452444, 452445, 452446, 452447, 452448, 452452, 452454, 452455, 452456, 452457, 452458, 452459, 452463, 452465, 452466, 452467, 452468, 452469, 452470, 452471, 452472, 452475, 452476, 452478, 452479, 452480, 452481, 452482, 452486, 452489, 452490, 452491, 452493, 452494, 452495, 452499, 452502, 452503, 452504, 452506, 452507, 452508, 452509, 452510, 452511, 452514, 452515, 452516, 452517, 452518, 452442, 452497, 452513, 452548

+++ src/gui/capture/capture.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020-2021  Marco Bortolin
+ * Copyright (C) 2020-2023  Marco Bortolin
  *
  * This file is part of IBMulator.
  *
@@ -182,7 +182,7 @@ void Capture::audio_sink(const std::vector<int16_t> &_data, int _category)
 		size_t data_len = _data.size() * 2;
 		size_t written = m_audio_buffer.write((uint8_t*)(&_data[0]), data_len);
 		if(written < data_len) {
-			PDEBUGF(LOG_V0, LOG_GUI, "Capture: audio buffer overrun: lost data: %d bytes\n", data_len-written);
+			PDEBUGF(LOG_V0, LOG_GUI, "Capture: audio buffer overrun: lost data: %zu bytes\n", size_t(data_len-written));
 		}
 	}
 }
